Change xdf_define_array API to be consistent with other libraries
In xdf_define_array, change the type of the strides argument (array of
unsigned int to array of const size_t) to be more consistent with the
other libraries

+++ src/xdffile.h
@@ -65,7 +65,7 @@ struct xdf {
 	unsigned int nbatch;
 	struct data_batch* batch;
 	unsigned int narrays;
-	unsigned int* array_stride;	
+	size_t* array_stride;	
 
 	struct xdfch* defaultch;
 	/* Data format specific behavior */
